Add files via upload
Hey fellas

So I was digging through and pulling together some code from my last run through of this class and I was able to get the parser program working based on the specs Professor Lewis provided on the args struct. It's a bit of an eye chart but I promise it works (for simple stuff anyway). At this point, I have the parser done and a bare bones version of the fork and exec pair. I hadn't seen the change directory function that Andrew's using, I'll have to play around with that tomorrow. Anyway, here's what I've got so far. More to come!

Lucas